feat: Add popular attractions display with auto-scroll feature
- Created a new HTML template for displaying popular attractions.
- Implemented a scrolling list of attractions in the TravelUI class.
- Enhanced the display_map method to handle map markers more efficiently.
- Updated the main.py to load popular attractions data from JSON.
- Added CSS styles for the background and scrolling effects.
- Introduced JavaScript for auto-scrolling functionality in the attractions section.
- Improved session management and error handling in the Flask app.
