+++ agents/chat_agent.py
@@ -1,7 +1,6 @@
 import os
 from langchain_openai import ChatOpenAI
-
-from langchain.schema import SystemMessage, HumanMessage, AIMessage
+from langchain_core.messages import SystemMessage, HumanMessage, AIMessage
 import json
 class ChatAgent:
     def __init__(self, model_name="gpt-3.5-turbo"):
@@ -27,25 +26,49 @@ class ChatAgent:
             self.conversation_history.append(self._init_system_message())
         
         # Merge new inputs into state
-        if user_input:
+        if user_input and user_input.strip():
             new_info = self.extract_info_from_message(user_input)
             for field, value in new_info.items():
                 if value:
                     state[field] = value
+                    print(f"Updated state: {field} = {value}")  # Debug log
         
-        # Add user input to conversation
-        self.conversation_history.append(HumanMessage(content=user_input))
+        # Add user input to conversation if not empty
+        if user_input and user_input.strip():
+            self.conversation_history.append(HumanMessage(content=user_input))
+        
+        # Get AI response based on current state and conversation history
+        messages = self.conversation_history.copy()
+        messages.append(SystemMessage(content=f"""
+        Current state: {json.dumps(state, ensure_ascii=False)}
+        Required fields: {json.dumps(self.required_fields, ensure_ascii=False)}
+        Missing fields: {json.dumps([f for f in self.required_fields if not state.get(f)], ensure_ascii=False)}
+        Please help the user complete the missing information in a natural way.
+        Remember to acknowledge information that has already been provided.
+        """))
+        
+        try:
+            response = self.model(messages)
+            self.conversation_history.append(AIMessage(content=response.content))
+        except Exception as e:
+            print(f"Error getting AI response: {e}")
+            response = AIMessage(content="抱歉，我遇到了一些问题。请再试一次。")
         
         missing = [f for f in self.required_fields if not state.get(f)]
         if missing:
-            prompt = (
-                "下面还有一些信息没给我，方便告诉我吗？\n" +
-                "\n".join(f"- {field}" for field in missing) +
-                "\n您可以一句话补充多个信息。"
-            )
-            return {"response": prompt, "missing_fields": missing, "complete": False}
+            return {
+                "response": response.content,
+                "missing_fields": missing,
+                "complete": False,
+                "state": state.copy()  # Return a copy of the state
+            }
         else:
-            return {"response": "👌 信息收集完毕，准备为您推荐景点。", "missing_fields": [], "complete": True}
+            return {
+                "response": response.content,
+                "missing_fields": [],
+                "complete": True,
+                "state": state.copy()  # Return a copy of the state
+            }
     
     def interact_with_user(self, message, state=None):
         """Process user message and generate a response"""
