+++ agents/communication_agent.py
@@ -1,5 +1,5 @@
-from langchain.chat_models import ChatOpenAI
-from langchain.schema import SystemMessage, HumanMessage
+from langchain_openai import ChatOpenAI
+from langchain_core.messages import SystemMessage, HumanMessage
 
 class CommunicationAgent:
     def __init__(self, model_name="gpt-3.5-turbo"):
