+++ data/popular_attractions.json
@@ -1,0 +1,37 @@
+[
+    {
+        "name": "Eiffel Tower",
+        "city": "Paris",
+        "country": "France",
+        "image_url": "https://images.unsplash.com/photo-1742756565106-2e0edbf37225?q=80&w=3087&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+        "description": "Iconic iron lattice tower."
+    },
+    {
+        "name": "Statue of Liberty",
+        "city": "New York",
+        "country": "USA",
+        "image_url": "https://plus.unsplash.com/premium_photo-1681803531285-75db948035d3?q=80&w=3087&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+        "description": "Symbol of freedom."
+    },
+    {
+        "name": "Great Wall",
+        "city": "Beijing",
+        "country": "China",
+        "image_url": "https://images.unsplash.com/photo-1640608486292-f52ac5d7cd76?q=80&w=3099&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+        "description": "Ancient defensive structure."
+    },
+    {
+        "name": "Taj Mahal",
+        "city": "Agra",
+        "country": "India",
+        "image_url": "https://images.unsplash.com/photo-1564507592333-c60657eea523",
+        "description": "Marble mausoleum."
+    },
+    {
+        "name": "Sydney Opera House",
+        "city": "Sydney",
+        "country": "Australia",
+        "image_url": "https://images.unsplash.com/photo-1606770951120-86e632c5bb52?q=80&w=2942&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+        "description": "Iconic performing arts center."
+    }
+]
