--- frontend/static/css/style.css
@@ -1,49 +1,0 @@
-.selected-marker {
-    background-color: #ff0000;
-    width: 20px;
-    height: 20px;
-    border-radius: 50%;
-    border: 2px solid white;
-    box-shadow: 0 0 5px rgba(0, 0, 0, 0.5);
-}
-
-.selected-marker-inner {
-    width: 10px;
-    height: 10px;
-    background-color: white;
-    border-radius: 50%;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-.attraction-item {
-    background-color: #f8f9fa;
-    padding: 10px;
-    margin-bottom: 10px;
-    border-radius: 5px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.attraction-item h4 {
-    margin: 0 0 5px 0;
-}
-
-.attraction-item p {
-    margin: 0 0 10px 0;
-    color: #666;
-}
-
-.attraction-item button {
-    background-color: #dc3545;
-    color: white;
-    border: none;
-    padding: 5px 10px;
-    border-radius: 3px;
-    cursor: pointer;
-}
-
-.attraction-item button:hover {
-    background-color: #c82333;
-}
