--- frontend/static/css/styles.css
@@ -1,239 +1,0 @@
-/* Main styles for Travel AI Assistant */
-
-/* Base Styles */
-body {
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    background-color: #f8f9fa;
-    color: #333;
-}
-
-/* Chat container */
-#chat-container {
-    height: 400px;
-    overflow-y: auto;
-    padding: 10px;
-    background-color: #f8f9fa;
-    border-radius: 5px;
-}
-
-/* Chat messages */
-.chat-message {
-    margin-bottom: 15px;
-    clear: both;
-    overflow: hidden;
-}
-
-.message-content {
-    padding: 10px 15px;
-    border-radius: 18px;
-    display: inline-block;
-    max-width: 80%;
-}
-
-.user {
-    text-align: right;
-}
-
-.user .message-content {
-    background-color: #007bff;
-    color: white;
-    float: right;
-}
-
-.assistant .message-content {
-    background-color: #e9ecef;
-    color: #212529;
-    float: left;
-}
-
-/* Map styles */
-#map {
-    border-radius: 5px;
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-}
-
-/* Card customization */
-.card {
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.05);
-    margin-bottom: 20px;
-}
-
-.card-header {
-    font-weight: 600;
-}
-
-/* Navbar customization */
-.navbar-brand {
-    font-weight: 600;
-    letter-spacing: 0.5px;
-}
-
-/* Button styling */
-.btn-outline-light:hover {
-    background-color: rgba(255, 255, 255, 0.1);
-}
-
-/* Itinerary styling */
-#itinerary-container {
-    max-height: 300px;
-    overflow-y: auto;
-}
-
-/* Recommendations styling */
-#recommendations-container {
-    max-height: 300px;
-    overflow-y: auto;
-}
-
-/* Loading spinner */
-#loading-spinner {
-    background-color: rgba(255, 255, 255, 0.7);
-    z-index: 1000;
-    border-radius: 8px;
-    padding: 20px;
-}
-
-/* Utility classes */
-.text-small {
-    font-size: 0.875rem;
-}
-
-/* Category markers on map */
-.map-marker {
-    border-radius: 50%;
-    border: 2px solid white;
-    text-align: center;
-    line-height: 30px;
-    font-weight: bold;
-}
-
-.marker-museum {
-    background-color: #3498db;
-    color: white;
-}
-
-.marker-landmark {
-    background-color: #e74c3c;
-    color: white;
-}
-
-.marker-nature {
-    background-color: #2ecc71;
-    color: white;
-}
-
-.marker-entertainment {
-    background-color: #9b59b6;
-    color: white;
-}
-
-.marker-shopping {
-    background-color: #f39c12;
-    color: white;
-}
-
-/* Price level indicators */
-.price-level {
-    color: #28a745;
-    font-weight: bold;
-}
-
-/* Rating stars */
-.rating {
-    color: #ffc107;
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-    #chat-container {
-        height: 300px;
-    }
-    
-    #map {
-        height: 250px !important;
-    }
-    
-    .message-content {
-        max-width: 90%;
-    }
-}
-
-/* Animation for chat messages */
-.chat-message {
-    animation: fadeIn 0.3s ease-in-out;
-}
-
-@keyframes fadeIn {
-    from { opacity: 0; transform: translateY(10px); }
-    to { opacity: 1; transform: translateY(0); }
-}
-
-/* Custom scrollbar */
-::-webkit-scrollbar {
-    width: 8px;
-}
-
-::-webkit-scrollbar-track {
-    background: #f1f1f1;
-    border-radius: 10px;
-}
-
-::-webkit-scrollbar-thumb {
-    background: #c1c1c1;
-    border-radius: 10px;
-}
-
-::-webkit-scrollbar-thumb:hover {
-    background: #a8a8a8;
-}
-
-.attraction-item {
-    background-color: #f8f9fa;
-    border-radius: 8px;
-    padding: 15px;
-    margin-bottom: 10px;
-    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-}
-
-.attraction-item h4 {
-    color: #333;
-    margin-bottom: 5px;
-}
-
-.attraction-item p {
-    color: #666;
-    margin-bottom: 10px;
-}
-
-.attraction-item button {
-    background-color: #dc3545;
-    color: white;
-    border: none;
-    padding: 5px 10px;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-.attraction-item button:hover {
-    background-color: #c82333;
-}
-
-.selected-marker {
-    background-color: #dc3545;
-    border: 2px solid white;
-    border-radius: 50%;
-    width: 20px;
-    height: 20px;
-    box-shadow: 0 0 5px rgba(0,0,0,0.3);
-}
-
-.selected-marker-inner {
-    background-color: white;
-    border-radius: 50%;
-    width: 8px;
-    height: 8px;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
