+++ frontend/templates/index.html
@@ -1,4 +1,3 @@
-/frontend/templates/index.html -->
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -66,7 +65,7 @@
                     <div class="card-body">
                         <form id="chat-form">
                             <div class="input-group">
-                                <input type="text" class="form-control" id="user-input" placeholder="Type your message here..." required>
+                                <input type="text" class="form-control" id="user-input" placeholder="Start your dream journey from here..." required>
                                 <button class="btn btn-primary" type="submit">
                                     <i class="fas fa-paper-plane"></i>
                                 </button>
@@ -75,12 +74,47 @@
                         <div class="mt-2 text-muted small">
                             <i class="fas fa-info-circle me-1"></i> Be specific about your travel preferences to get better recommendations.
                         </div>
+                        <!-- Missing fields alert -->
+                        <div id="missing-fields" class="alert alert-warning mt-3 d-none"></div>
                     </div>
                 </div>
             </div>
             
             <!-- Right Panel: Map and Results -->
             <div class="col-md-6">
+                <!-- Top World Attractions -->
+                <div class="card mb-3">
+                    <div class="card-header bg-primary text-white">
+                        <i class="fas fa-star me-2"></i> Top World Attractions
+                    </div>
+                    <div class="card-body p-0">
+                        <div class="scroll-container">
+                            <div class="scroll-content">
+                                {% for attraction in popular_attractions %}
+                                <div class="attraction-card" aria-label="Attraction: {{ attraction.name }}, {{ attraction.city }}, {{ attraction.country }}">
+                                    <img src="{{ attraction.image_url }}" alt="{{ attraction.name }}">
+                                    <div class="attraction-info">
+                                        <h4>{{ loop.index }}. {{ attraction.name }}</h4>
+                                        <p>{{ attraction.city }}, {{ attraction.country }}</p>
+                                        <p>{{ attraction.description }}</p>
+                                    </div>
+                                </div>
+                                {% endfor %}
+                                {% for attraction in popular_attractions %}
+                                <div class="attraction-card" aria-label="Attraction: {{ attraction.name }}, {{ attraction.city }}, {{ attraction.country }}">
+                                    <img src="{{ attraction.image_url }}" alt="{{ attraction.name }}">
+                                    <div class="attraction-info">
+                                        <h4>{{ loop.index }}. {{ attraction.name }}</h4>
+                                        <p>{{ attraction.city }}, {{ attraction.country }}</p>
+                                        <p>{{ attraction.description }}</p>
+                                    </div>
+                                </div>
+                                {% endfor %}
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                
                 <!-- Map -->
                 <div class="card mb-3">
                     <div class="card-header bg-primary text-white">
@@ -137,6 +171,5 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"></script>
     <script src="{{ url_for('static', filename='js/main.js') }}"></script>
-    <script src="{{ url_for('static', filename='tests/map_test.js') }}"></script>
 </body>
 </html>
