+++ frontend/templates/popular_attractions.html
@@ -1,0 +1,7 @@
+<!-- templates/popular_attractions.html -->
+<div class="scroll-container">
+    <div class="scroll-content">
+        {attractions_html}
+        {attractions_html} <!-- Duplicate content for seamless looping -->
+    </div>
+</div>
